moved development image target into the ci/runner/Dockerfile; refactored common get_image_full_name() function to allow for optional values for the server and tag prefix/suffix; get_image_full_name now requires a positional argument

+++ scripts/devel.sh
@@ -17,35 +17,39 @@
 # NOTE: this script should be exectued from the root of the mrc repo
 # ./scripts/devel.sh
 
-# command used on the local system to run docker, e.g. "docker", "nvidia-docker", etc.
-DOCKER_CMD=${DOCKER_CMD:-"nvidia-docker"}
+source "ci/runner/common.sh"
 
-# options to select which gpu devices to map into the container
-DOCKER_GPU_OPTS=${DOCKER_CMD:-" "}
+export DOCKER_REGISTRY_SERVER=""
+export DOCKER_REGISTRY_PATH="mrc"
+export DOCKER_TARGET="dev"
+export DOCKER_TAG_PREFIX=""
+export DOCKER_TAG_SUFFIX="$(date +'%y%m%d')"
 
-# extract the ci image name from the github actions definitions
-CI_IMAGE=$(yq '.jobs.ci_pipe.with.test_container' < .github/workflows/pull_request.yml)
+export SKIP_PUSH=yes
+SKIP_RUN=${SKIP_RUN:-""}
 
-# local image name and tag
-IMAGE_NAME=${IMAGE_NAME:-"mrc"}
-IMAGE_TAG=${TAG:-"dev-$(date +'%y%m%d')"}
+IMAGE_NAME=$(get_image_full_name ${DOCKER_TARGET})
+CONTAINER_NAME=$(echo ${IMAGE_NAME} | sed -e 's/:/-/g')
 
-set -e
+./ci/runner/build_and_push.sh
 
-docker build --build-arg FROM_IMAGE=${CI_IMAGE} --network=host -t ${IMAGE_NAME}:${IMAGE_TAG} -f scripts/Dockerfile .
+DOCKER_CMD=${DOCKER_CMD:-"docker"}
+DOCKER_GPU_OPTS=${DOCKER_GPU_OPTS:-"--gpus=all"}
 
-${DOCKER_CMD} run \
+if [ "${SKIP_RUN}" == "" ]; then
+  ${DOCKER_CMD} run \
     ${DOCKER_GPU_OPTS} \
     --rm -d \
     -v $PWD:/work \
     -v /cores:/cores \
     -v /cache:/cache \
     --workdir /work \
-    --name ${IMAGE_NAME}_${IMAGE_TAG} \
+    --name ${CONTAINER_NAME} \
     --net host \
     --ulimit core=-1 \
     --cap-add=SYS_PTRACE \
     --cap-add=SYS_ADMIN \
     --cap-add=SYS_NICE \
-    ${IMAGE_NAME}:${IMAGE_TAG} \
+    ${IMAGE_NAME} \
     sleep 999999999999999999999
+fi
